# specs/006-registration-requests-approval/contracts/registration-requests.openapi.yaml
openapi: 3.1.0
info:
  title: New Talents Registration Requests API
  version: 0.1.0
  description: Feature 006 only. Ordinary JSON never includes document bytes or raw private identity values.
  license: {name: Proprietary, identifier: LicenseRef-Proprietary}
servers:
  - url: /api
...
paths:
  /registration-requests/validate-identity:
    post:
      tags: [Public registration]
      operationId: validatePublicRegistrationIdentity
      summary: Validate exact identifiers for one bounded public registration step
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IdentityConflictValidation'}}}}
      responses: {'200': {description: Identifiers available}, '422': {$ref: '#/components/responses/ValidationError'}}
  /academies/{academyId}/registration-requests/validate-identity:
    post:
      tags: [Academy]
      operationId: validateAcademyRegistrationIdentity
      summary: Validate exact identifiers in an authorized academy context
      parameters: [{$ref: '#/components/parameters/AcademyId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/IdentityConflictValidation'}}}}
      responses: {'200': {description: Identifiers available}, '404': {$ref: '#/components/responses/SafeDenied'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/personal-adult:
    post:
      tags: [Public registration]
      operationId: createPersonalAdultRequest
      summary: Create a personal adult registration draft
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/PersonalAdultCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '409': {$ref: '#/components/responses/SafeConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/represented-minor:
    post:
      tags: [Public registration]
      operationId: createRepresentedMinorRequest
      summary: Create a represented minor registration draft
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RepresentedMinorCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '409': {$ref: '#/components/responses/SafeConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/academies/formal:
    post:
      tags: [Public registration]
      operationId: createFormalAcademyRequest
      summary: Create a formal academy registration draft
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/FormalAcademyCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '409': {$ref: '#/components/responses/SafeConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/academies/natural-person:
    post:
      tags: [Public registration]
      operationId: createNaturalPersonAcademyRequest
      summary: Create a natural-person academy registration draft
      security: []
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/NaturalPersonAcademyCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '409': {$ref: '#/components/responses/SafeConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /academies/{academyId}/registration-requests/accounts:
    post:
      tags: [Academy]
      operationId: createAdditionalAcademyAccountRequest
      summary: Create an additional academy account request
      parameters: [{$ref: '#/components/parameters/AcademyId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AdditionalAcademyAccountCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '403': {$ref: '#/components/responses/SafeDenied'}, '409': {$ref: '#/components/responses/SafeConflict'}}
  /academies/{academyId}/registration-requests/players/adult:
    post:
      tags: [Academy]
      operationId: createAcademyAdultPlayerRequest
      summary: Create an academy adult player request
      parameters: [{$ref: '#/components/parameters/AcademyId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AcademyAdultPlayerCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '403': {$ref: '#/components/responses/SafeDenied'}, '409': {$ref: '#/components/responses/SafeConflict'}}
  /academies/{academyId}/registration-requests/players/minor:
    post:
      tags: [Academy]
      operationId: createAcademyMinorPlayerRequest
      summary: Create an academy minor player request
      parameters: [{$ref: '#/components/parameters/AcademyId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/AcademyMinorPlayerCreate'}}}}
      responses: {'201': {$ref: '#/components/responses/RequestCreated'}, '403': {$ref: '#/components/responses/SafeDenied'}, '409': {$ref: '#/components/responses/SafeConflict'}}
  /registration-requests/{requestId}:
    get:
      tags: [Applicant]
      operationId: getOwnRegistrationRequest
      summary: Read an authorized owned request
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      responses: {'200': {$ref: '#/components/responses/RequestRead'}, '404': {$ref: '#/components/responses/SafeNotFound'}}
    patch:
      tags: [Applicant]
      operationId: updateOwnRegistrationRequest
      summary: Update an editable owned request
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {required: true, content: {application/json: {schema: {$ref: '#/components/schemas/RequestUpdate'}}}}
      responses: {'200': {$ref: '#/components/responses/RequestRead'}, '409': {$ref: '#/components/responses/VersionConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/{requestId}/submit:
    post:
      tags: [Applicant]
      operationId: submitRegistrationRequest
      summary: Submit a complete registration request
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {$ref: '#/components/requestBodies/Transition'}
      responses: {'200': {$ref: '#/components/responses/RequestRead'}, '409': {$ref: '#/components/responses/VersionConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/{requestId}/resubmit:
    post:
      tags: [Applicant]
      operationId: resubmitRegistrationRequest
      summary: Resubmit a corrected registration request
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody: {$ref: '#/components/requestBodies/Transition'}
      responses: {'200': {$ref: '#/components/responses/RequestRead'}, '409': {$ref: '#/components/responses/VersionConflict'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/{requestId}/evidence:
    post:
      tags: [Applicant]
      operationId: uploadRegistrationEvidence
      summary: Upload temporary evidence to quarantine
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [category, expectedVersion, file]
              properties:
                category: {$ref: '#/components/schemas/EvidenceCategory'}
                expectedVersion: {type: integer, minimum: 0}
                file: {type: string, format: binary}
      responses: {'201': {description: Evidence metadata, content: {application/json: {schema: {$ref: '#/components/schemas/EvidenceEnvelope'}}}}, '413': {$ref: '#/components/responses/UploadTooLarge'}, '415': {$ref: '#/components/responses/UnsupportedEvidence'}, '422': {$ref: '#/components/responses/ValidationError'}}
  /registration-requests/{requestId}/evidence/{evidenceId}:
    delete:
      tags: [Applicant]
      operationId: replaceOrRemoveRegistrationEvidence
      summary: Retire current evidence from an editable request
      parameters: [{$ref: '#/components/parameters/RequestId'}, {$ref: '#/components/parameters/EvidenceId'}, {$ref: '#/components/parameters/ExpectedVersion'}]
      responses: {'204': {description: Evidence retired and deletion scheduled}, '409': {$ref: '#/components/responses/VersionConflict'}, '404': {$ref: '#/components/responses/SafeNotFound'}}
  /registration-requests/{requestId}/deletion:
    get:
      tags: [Applicant]
      operationId: getOwnEvidenceDeletionStatus
      summary: Read safe evidence deletion status
      parameters: [{$ref: '#/components/parameters/RequestId'}]
      responses: {'200': {description: Safe deletion status, content: {application/json: {schema: {$ref: '#/components/schemas/DeletionEnvelope'}}}}, '404': {$ref: '#/components/responses/SafeNotFound'}}
  /academies/{academyId}/registration-requests:
    get:
      tags: [Academy]
      operationId: listAcademyRegistrationRequests
      summary: List requests in an authorized academy context
      parameters: [{$ref: '#/components/parameters/AcademyId'}, {$ref: '#/components/parameters/Cursor'}, {$ref: '#/components/parameters/PageSize'}]
      responses: {'200': {$ref: '#/components/responses/RequestPage'}, '403': {$ref: '#/components/responses/SafeDenied'}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    RequestId: {name: requestId, in: path, required: true, schema: {type: string, format: uuid}}
    EvidenceId: {name: evidenceId, in: path, required: true, schema: {type: string, format: uuid}}
    AcademyId: {name: academyId, in: path, required: true, schema: {type: string, format: uuid}}
    ExpectedVersion: {name: expectedVersion, in: query, required: true, schema: {type: integer, minimum: 0}}
    Cursor: {name: cursor, in: query, schema: {type: string, maxLength: 512}}
    PageSize: {name: limit, in: query, schema: {type: integer, minimum: 1, maximum: 50, default: 20}}
  requestBodies:
    Transition:
      required: true
      content: {application/json: {schema: {$ref: '#/components/schemas/TransitionCommand'}}}
  responses:
    RequestCreated: {description: Restricted request created, content: {application/json: {schema: {$ref: '#/components/schemas/RequestEnvelope'}}}}
    RequestRead: {description: Authorized request projection, content: {application/json: {schema: {$ref: '#/components/schemas/RequestEnvelope'}}}}
    RequestPage: {description: Cursor page, content: {application/json: {schema: {$ref: '#/components/schemas/RequestPage'}}}}
    SafeDenied: {description: Safely denied, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    SafeNotFound: {description: Missing or not authorized without existence disclosure, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    SafeConflict: {description: Non-enumerating registration conflict, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    VersionConflict: {description: Stale lifecycle version, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    ValidationError: {description: Safe field/category validation, content: {application/json: {schema: {$ref: '#/components/schemas/ValidationErrorEnvelope'}}}}
    UploadTooLarge: {description: Configured size exceeded, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
    UnsupportedEvidence: {description: Unsupported or mismatched content, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorEnvelope'}}}}
  schemas:
    IdentityConflictValidation:
      type: object
      additionalProperties: false
      required: [requestType, checks]
      properties:
        requestType: {type: string, enum: [PERSONAL_ADULT, REPRESENTED_MINOR, FORMAL_ACADEMY, NATURAL_PERSON_ACADEMY, ADDITIONAL_ACADEMY_ACCOUNT, ACADEMY_ADULT_PLAYER, ACADEMY_MINOR_PLAYER]}
        checks:
          type: array
          minItems: 1
          maxItems: 3
          items:
            oneOf:
              - {type: object, additionalProperties: false, required: [field, documentType, documentNumber], properties: {field: {type: string, enum: [person.documentNumber, representative.documentNumber, minor.documentNumber, academy.responsiblePerson.documentNumber, player.documentNumber]}, documentType: {type: string, enum: [CC, TI, RC, CE, PASSPORT]}, documentNumber: {type: string, minLength: 1, maxLength: 80}}}
              - {type: object, additionalProperties: false, required: [field, nit], properties: {field: {const: nit}, nit: {type: string, minLength: 1, maxLength: 40}}}
              - {type: object, additionalProperties: false, required: [field, academyName], properties: {field: {const: academy.academyName}, academyName: {type: string, minLength: 1, maxLength: 180}}}
...
    EvidenceStatus: {type: string, enum: [QUARANTINED, SCANNING, CLEAN, REJECTED, REPLACED, DELETION_PENDING, DELETED]}
    DeletionStatus: {type: string, enum: [PENDING, IN_PROGRESS, COMPLETED, RECOVERY_REQUIRED]}
    EvidenceCategory: {type: string, enum: [IDENTITY_FRONT, IDENTITY_BACK, MINOR_CIVIL_IDENTITY, REPRESENTATION_AUTHORITY, RUT, EXISTENCE_CERTIFICATE, RESPONSIBLE_AUTHORITY, OPERATION_PROOF, ADULT_AUTHORIZATION, ACADEMY_ACCOUNT_AUTHORIZATION]}
    CredentialInput:
      type: object
      additionalProperties: false
      required: [email, password, passwordConfirmation]
      properties: {email: {type: string, format: email, maxLength: 254}, password: {type: string, format: password, minLength: 12, maxLength: 128, writeOnly: true}, passwordConfirmation: {type: string, format: password, minLength: 12, maxLength: 128, writeOnly: true}}
    PrivatePersonInput:
      type: object
      additionalProperties: false
      required: [legalNames, legalSurnames, documentType, documentNumber, birthDate, country, city]
      properties: {legalNames: {type: string, minLength: 1, maxLength: 120}, legalSurnames: {type: string, minLength: 1, maxLength: 120}, documentType: {type: string, enum: [CC, TI, RC, CE, PASSPORT]}, documentNumber: {type: string, minLength: 1, maxLength: 80, description: "Validated according to documentType; legacy Colombian CC numbers remain valid."}, birthDate: {type: string, format: date, description: "Exact YYYY-MM-DD; impossible and future dates are rejected."}, country: {type: string, enum: [CO]}, city: {type: string, pattern: '^[0-9]{5}$', description: "Stable DANE DIVIPOLA municipality code."}, phone: {type: string, minLength: 1, maxLength: 40}}
    MinorInput: {$ref: '#/components/schemas/PrivatePersonInput'}
    ConsentInput:
      type: object
      additionalProperties: false
      required: [privacyVersion, privacyAccepted, truthfulnessAccepted]
      properties: {privacyVersion: {type: string, maxLength: 40}, privacyAccepted: {type: boolean, const: true}, truthfulnessAccepted: {type: boolean, const: true}, representationAccepted: {type: boolean}, minorTreatmentAccepted: {type: boolean}, academyPresentationAccepted: {type: boolean}}
    PersonalAdultCreate:
      type: object
      additionalProperties: false
      required: [credentials, person, actingForSelf, consent]
      properties: {credentials: {$ref: '#/components/schemas/CredentialInput'}, person: {$ref: '#/components/schemas/PrivatePersonInput'}, actingForSelf: {type: boolean, const: true}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    RepresentedMinorCreate:
      type: object
      additionalProperties: false
      required: [credentials, representative, minor, relationship, authorityDeclared, consent]
      properties: {credentials: {$ref: '#/components/schemas/CredentialInput'}, representative: {$ref: '#/components/schemas/PrivatePersonInput'}, minor: {$ref: '#/components/schemas/MinorInput'}, relationship: {type: string, enum: [MOTHER, FATHER, LEGAL_GUARDIAN]}, authorityDeclared: {type: boolean, const: true}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    AcademyBaseInput:
      type: object
      additionalProperties: false
      required: [academyName, country, city, responsiblePerson]
      properties: {academyName: {type: string, minLength: 1, maxLength: 180}, country: {type: string, enum: [CO]}, city: {type: string, pattern: '^[0-9]{5}$', description: "Stable DANE DIVIPOLA municipality code."}, trainingPlace: {type: string, minLength: 1, maxLength: 180}, responsiblePerson: {$ref: '#/components/schemas/PrivatePersonInput'}}
    FormalAcademyCreate:
      type: object
      additionalProperties: false
      required: [credentials, academy, organizationType, nit, authorityDeclared, consent]
      properties: {credentials: {$ref: '#/components/schemas/CredentialInput'}, academy: {$ref: '#/components/schemas/AcademyBaseInput'}, organizationType: {type: string, maxLength: 80}, nit: {type: string, maxLength: 40}, authorityDeclared: {type: boolean, const: true}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    NaturalPersonAcademyCreate:
      type: object
      additionalProperties: false
      required: [credentials, academy, operationDeclared, proofCategories, consent]
      properties: {credentials: {$ref: '#/components/schemas/CredentialInput'}, academy: {$ref: '#/components/schemas/AcademyBaseInput'}, operationDeclared: {type: boolean, const: true}, proofCategories: {type: array, minItems: 1, uniqueItems: true, items: {type: string, enum: [RUT, MUNICIPAL_OR_SPORT_CERTIFICATION, PLACE_USE_AUTHORIZATION, OPERATION_CONTRACT_OR_REGISTER, OTHER_CONTROLLED]}}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    AdditionalAcademyAccountCreate:
      type: object
      additionalProperties: false
      required: [person, function, responsibleAuthorization, consent]
      properties: {person: {$ref: '#/components/schemas/PrivatePersonInput'}, function: {type: string, maxLength: 120}, responsibleAuthorization: {type: boolean, const: true}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    AcademyAdultPlayerCreate:
      type: object
      additionalProperties: false
      required: [player, adultAuthorization, consent]
      properties: {player: {$ref: '#/components/schemas/PrivatePersonInput'}, adultAuthorization: {type: boolean, const: true}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    AcademyMinorPlayerCreate:
      type: object
      additionalProperties: false
      required: [minor, representative, relationship, authorityDeclared, consent]
      properties: {minor: {$ref: '#/components/schemas/MinorInput'}, representative: {$ref: '#/components/schemas/PrivatePersonInput'}, relationship: {type: string, enum: [MOTHER, FATHER, LEGAL_GUARDIAN]}, authorityDeclared: {type: boolean, const: true}, consent: {$ref: '#/components/schemas/ConsentInput'}}
    RequestUpdate:
      type: object
      additionalProperties: false
      required: [expectedVersion, details]
      properties: {expectedVersion: {type: integer, minimum: 0}, details: {oneOf: [{$ref: '#/components/schemas/PersonalAdultCreate'}, {$ref: '#/components/schemas/RepresentedMinorCreate'}, {$ref: '#/components/schemas/FormalAcademyCreate'}, {$ref: '#/components/schemas/NaturalPersonAcademyCreate'}, {$ref: '#/components/schemas/AdditionalAcademyAccountCreate'}, {$ref: '#/components/schemas/AcademyAdultPlayerCreate'}, {$ref: '#/components/schemas/AcademyMinorPlayerCreate'}]}}
    TransitionCommand: {type: object, additionalProperties: false, required: [expectedVersion, idempotencyKey], properties: {expectedVersion: {type: integer, minimum: 0}, idempotencyKey: {type: string, format: uuid}}}
...
    EvidenceMetadata: {type: object, additionalProperties: false, required: [id, category, status, sizeBytes], properties: {id: {type: string, format: uuid}, category: {$ref: '#/components/schemas/EvidenceCategory'}, status: {$ref: '#/components/schemas/EvidenceStatus'}, sizeBytes: {type: integer, minimum: 0}, correctionRequired: {type: boolean}}}
    RequestSummary: {type: object, additionalProperties: false, required: [id, type, status, version, capabilities, createdAt], properties: {id: {type: string, format: uuid}, type: {$ref: '#/components/schemas/RegistrationRequestType'}, status: {$ref: '#/components/schemas/RegistrationRequestStatus'}, version: {type: integer, minimum: 0}, safeApplicantLabel: {type: string}, academyLabel: {type: string}, submittedAt: {type: string, format: date-time}, createdAt: {type: string, format: date-time}, evidenceComplete: {type: boolean}, correctionRequired: {type: boolean}, safeReason: {type: string}, capabilities: {type: array, uniqueItems: true, items: {type: string}}}}
    RequestDetail: {allOf: [{$ref: '#/components/schemas/RequestSummary'}, {type: object, properties: {evidence: {type: array, items: {$ref: '#/components/schemas/EvidenceMetadata'}}, deletion: {$ref: '#/components/schemas/DeletionStatusView'}}}]}
...
    DeletionStatusView: {type: object, additionalProperties: false, required: [status, totalItems, completedItems], properties: {status: {$ref: '#/components/schemas/DeletionStatus'}, totalItems: {type: integer, minimum: 0}, completedItems: {type: integer, minimum: 0}, lastUpdatedAt: {type: string, format: date-time}}}
...
    RequestEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/RequestDetail'}}}
...
    EvidenceEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/EvidenceMetadata'}}}
    DeletionEnvelope: {type: object, additionalProperties: false, required: [data], properties: {data: {$ref: '#/components/schemas/DeletionStatusView'}}}
...
    RequestPage: {type: object, additionalProperties: false, required: [data, pagination], properties: {data: {type: array, items: {$ref: '#/components/schemas/RequestSummary'}}, pagination: {$ref: '#/components/schemas/Pagination'}}}
    ErrorEnvelope: {type: object, additionalProperties: false, required: [error], properties: {error: {type: object, additionalProperties: false, required: [code, message], properties: {code: {type: string}, message: {type: string}, correlationId: {type: string}}}}}
    ValidationIssue: {type: object, additionalProperties: false, required: [field, code], properties: {field: {type: string}, code: {type: string}, message: {type: string}}}
    ValidationErrorEnvelope: {allOf: [{$ref: '#/components/schemas/ErrorEnvelope'}, {type: object, properties: {issues: {type: array, items: {$ref: '#/components/schemas/ValidationIssue'}}}}]}
security:
  - bearerAuth: []